chore: init rootRoute && pages config
