+++ packages/varlet-cli/site/index.html
@@ -20,6 +20,10 @@
       font-style: normal;
       font-display: swap;
     }
+    *{
+      margin: 0px;
+      padding: 0px;
+    }
   </style>
   <script>
     if ('<%= baidu %>') {
