+++ packages/varlet-cli/src/compiler/compileSiteEntry.ts
@@ -4,20 +4,23 @@ import {
   EXAMPLE_DIR_INDEX,
   EXAMPLE_DIR_NAME,
   ROOT_DOCS_DIR,
+  ROOT_PAGES_DIR,
   SITE,
   SITE_DIR,
   SITE_MOBILE_ROUTES,
+  SITE_PC_DIR,
   SITE_PC_ROUTES,
   SRC_DIR,
 } from '../shared/constant'
 import { copy, pathExistsSync, readdir, readdirSync } from 'fs-extra'
 import { resolve } from 'path'
-import { isMD, outputFileSyncOnChange } from '../shared/fsUtils'
+import { isMD, isSFC, outputFileSyncOnChange } from '../shared/fsUtils'
 import { getVarletConfig } from '../config/varlet.config'
 
 const EXAMPLE_COMPONENT_NAME_RE = /\/([-\w]+)\/example\/index.vue/
 const COMPONENT_DOCS_RE = /\/([-\w]+)\/docs\/([-\w]+)\.md/
 const ROOT_DOCS_RE = /\/docs\/([-\w]+)\.([-\w]+)\.md/
+const ROOT_PAGE_RE = /\/pages\/([-\w]+)\.([-\w]+)\.vue/
 
 export function getExampleRoutePath(examplePath: string): string {
   return '/' + examplePath.match(EXAMPLE_COMPONENT_NAME_RE)?.[1]
@@ -33,6 +36,27 @@ export function getRootDocsRoutePath(rootDocsPath: string): string {
   const [, routePath, language] = rootDocsPath.match(ROOT_DOCS_RE) ?? []
 
   return `/${language}/${routePath}`
+}
+
+export function getRootPagesPath(rootPagePath: string): string {
+  const [, routePath, language] = rootPagePath.match(ROOT_PAGE_RE) ?? []
+  return `/${language}/${routePath}`
+}
+
+export async function findRootPagesPath(): Promise<string[]> {
+  if (!pathExistsSync(ROOT_PAGES_DIR)) {
+    return []
+  }
+
+  const dir: string[] = await readdir(ROOT_PAGES_DIR)
+
+  const buildPath = (filename: string) => resolve(ROOT_PAGES_DIR, filename)
+
+  const existPath = (filename: string) => isSFC(buildPath(filename))
+
+  const slashPath = (filename: string) => slash(buildPath(filename))
+
+  return dir.filter(existPath).map(slashPath)
 }
 
 export async function findExamplePaths(): Promise<string[]> {
@@ -105,7 +129,15 @@ export async function buildMobileSiteRoutes() {
 }
 
 export async function buildPcSiteRoutes() {
-  const [componentDocsPaths, rootDocsPaths] = await Promise.all([findComponentDocsPaths(), findRootDocsPaths()])
+  const rootPagesRoutes = rootPagesPaths.map(
+    (rootPagesPath) => `
+  {
+    path: '${getRootPagesPath(rootPagesPath)}',
+    // @ts-ignore
+    component: () => import('${rootPagesPath}')
+  }\
+`
+  )
 
   const componentDocsRoutes = componentDocsPaths.map(
     (componentDocsPath) => `
@@ -127,10 +159,21 @@ export async function buildPcSiteRoutes() {
 `
   )
 
+  const layoutRoutes = `
+    {
+      path: '/layout',
+      // @ts-ignore
+      component:()=> import('${SITE_PC_DIR}/Layout.vue'),
+      children: [
+        ${[...componentDocsRoutes, rootDocsRoutes].join(',')}, 
+      ]
+    }\
+  `
   const source = `export default [\
-  ${[...componentDocsRoutes, rootDocsRoutes].join(',')}
-]`
-
+    ${rootPagesRoutes.join(',')},
+    \
+    ${layoutRoutes}
+  ]`
   outputFileSyncOnChange(SITE_PC_ROUTES, source)
 }
