+++ packages/varlet-cli/varlet.default.config.js
@@ -19,7 +19,7 @@ module.exports = {
   useMobile: false,
   pc: {
     menu: [],
-    redirect: '/home',
+    redirect: '/rootRoute',
     title: {
       'zh-CN': '面向 Vue3 的 Material 风格移动端组件库',
       'en-US': 'Material design mobile components built for Vue3',
