+++ packages/varlet-ui/module.d.ts
@@ -1,0 +1,2 @@
+declare module '@config'
+declare module 'vue-router'
