+++ packages/varlet-ui/varlet.config.js
@@ -4,6 +4,7 @@ module.exports = {
     baidu: 'https://hm.baidu.com/hm.js?5c628ce58967c90ff4dd9c8803d930fa',
   },
   pc: {
+    redirce: '/rootRoute',
     menu: [
       {
         text: {
